% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{aa_changes_phenotype}
\alias{aa_changes_phenotype}
\title{aa_changes_phenotype}
\usage{
aa_changes_phenotype(samples, product_id, posns, alts, gene, phenotype)
}
\arguments{
\item{samples}{(number) sample_id}

\item{product_id}{(number) product_id}

\item{posns}{(number) codon posiiotn in gene where changes lead to rsistance}

\item{alts}{(character) amino acid changes in gene where changes lead to redsistance}

\item{gene}{(string)}

\item{phenotype(string)}{}
}
\value{
finds
}
\description{
aa_changes_phenotype
}
\examples{
gyrA_cipro_results <- aa_changes_phenotype(s_list,815,gyrA_posns,gyrA_alts,"gyrA","Ciprofloxacin")
}
