% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{antibiotic_data}
\alias{antibiotic_data}
\title{antibiotic_data}
\usage{
antibiotic_data(no_SNPs = TRUE)
}
\arguments{
\item{no_SNPs}{}
}
\value{
Dataframe of staphopia SNPs associated with known AR or (no_SNP=F) dataframe of antibiotic resistance loci or
}
\description{
antibiotic_data
}
\examples{
AR_data <- antibiotic_data()
}

