% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plots.R
\name{bonferroni_correction}
\alias{bonferroni_correction}
\title{bonferroni_correction}
\usage{
bonferroni_correction(alpha, n)
}
\arguments{
\item{alpha}{A significant p-value threshold}

\item{n}{Total number of independent tests.}
}
\value{
A -log10 float value.
}
\description{
Returns a Bonferroni corrected alpha value..
}
\examples{
bonferroni_correction(0.05, 100)
}
