% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{build_url}
\alias{build_url}
\title{build_url}
\usage{
build_url(request)
}
\arguments{
\item{request}{}
}
\value{
Full URL endpoint.
}
\description{
Builds the endpoint url to query the API. This function should not be
directly called by the user.
}

