% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{cgmlst_distance}
\alias{cgmlst_distance}
\title{cgmlst_distance}
\usage{
cgmlst_distance(df)
}
\arguments{
\item{cgmlst_data_frame}{Data Frame from 'get_cgmlst' function}
}
\value{
Parsed JSON response.
}
\description{
Using output from 'get_cgmlst' determine the pairwise hamming distance between each sample.
}
\examples{
cgmlst_distance(cgmlst_data_frame)
}
