% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{create_snp_matrix}
\alias{create_snp_matrix}
\title{create_snp_matrix}
\usage{
create_snp_matrix(snps, samples, snp_info)
}
\arguments{
\item{snps}{A data frame returned from 'get_snps_by_samples'
'}

\item{samples}{A data frame of samples including sample_id and sample_tag}

\item{snp_info}{A data frame returned from 'get_snps_in_bulk'
'}
}
\value{
Parsed JSON response.
}
\description{
Given a data frame with 'sample_id' and 'snp_id' return an SNP matrix that
can be used conduct a GWAS.
}
\examples{
create_snp_matrix(snps, snp_info)
}
