% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{find_SNPs_by_gene}
\alias{find_SNPs_by_gene}
\title{find_SNPs_by_gene}
\usage{
find_SNPs_by_gene(coords, posns, alts, gene, phenotype)
}
\arguments{
\item{coords}{.}

\item{posns}{.}

\item{alts}{.}

\item{gene}{.}

\item{phenotype}{.}
}
\value{
dataframe of SNPs associated with AR gene
}
\description{
find_SNPs_by_gene
}

