% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_id}
\alias{format_id}
\title{format_id}
\usage{
format_id(x)
}
\arguments{
\item{x}{A numeric representation of an ID}
}
\value{
ID as an integer.
}
\description{
In case scientific notation is turned on, convert it to an integer. This
function should not be directly used by the user.
}
