% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ids}
\alias{format_ids}
\title{format_ids}
\usage{
format_ids(x)
}
\arguments{
\item{x}{A vector numeric representation of IDs}
}
\value{
ID as an integer.
}
\description{
In case scientific notation is turned on, convert all ids to an integer. This
function should not be directly used by the user.
}
