% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{get_assembly}
\alias{get_assembly}
\title{get_assembly}
\usage{
get_assembly(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all assembled contigs for a given sample.
}
\examples{
get_assembly(500)
}
