% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_assembly_stats}
\alias{get_assembly_stats}
\title{get_assembly_stats}
\usage{
get_assembly_stats(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all stats associated with an assembly for a given sample.
}
\examples{
get_assembly_stats(500)
}
