% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_assembly_stats_by_year}
\alias{get_assembly_stats_by_year}
\title{get_assembly_stats_by_year}
\usage{
get_assembly_stats_by_year(is_scaffolds = FALSE, is_plasmids = FALSE)
}
\arguments{
\item{is_scaffolds}{Boolean}

\item{is_plasmids}{Boolean}
}
\value{
Parsed JSON response.
}
\description{
Retrieve assembled stats by year based on when samples were first made public.
}
\examples{
get_assembly_stats_by_year(is_scaffolds=TRUE, is_plasmids=FALSE)
}
