% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{get_citations}
\alias{get_citations}
\title{get_citations}
\usage{
get_citations(pmid = FALSE)
}
\arguments{
\item{pmid}{An integer PubMed ID, or vector of PubMed IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all publications in a given PubMed ID(s).
}
\examples{
get_citations()
get_citations(22674929)
get_citations(c(22674929,26672018))
}
