% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_contig}
\alias{get_contig}
\title{get_contig}
\usage{
get_contig(contig_id)
}
\arguments{
\item{contig_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve assembled contigs by a given contig_id.  REWORK THIS QUERY
}
\examples{
get_contig(500)
}
