% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_contigs}
\alias{get_contigs}
\title{get_contigs}
\usage{
get_contigs(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all assembled contigs for a given sample.
}
\examples{
get_contigs(500)
}

