% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_gene_cluster}
\alias{get_gene_cluster}
\title{get_gene_cluster}
\usage{
get_gene_cluster(cluster_id)
}
\value{
Parsed JSON response.
}
\description{
Retrieve gene cluster for a given cluster_id.
}
\examples{
get_gene_cluster(500)
}
