% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{get_gene_cluster_counts}
\alias{get_gene_cluster_counts}
\title{get_gene_cluster_counts}
\usage{
get_gene_cluster_counts(sample_ids)
}
\value{
Parsed JSON response.
}
\description{
Retrieve gene cluster counts for a set of sample ids.
}
\examples{
get_gene_cluster_counts(500:505)
}
