% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{get_gene_clusters}
\alias{get_gene_clusters}
\title{get_gene_clusters}
\usage{
get_gene_clusters(sample_ids)
}
\value{
Parsed JSON response.
}
\description{
Retrieve gene cluster for a given cluster_id.
}
\examples{
get_gene_clusters(500:505)
}

