% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_gene_products}
\alias{get_gene_products}
\title{get_gene_products}
\usage{
get_gene_products(query = NULL)
}
\arguments{
\item{query}{An interger, query string or empty value.}
}
\value{
Parsed JSON response.
}
\description{
Retrieve a single gene product, gene products associated with a term, or
all gene products in Staphopia
}
\examples{
get_gene_products()
get_gene_products('meca')
get_gene_products(5:50)
}
