% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_genes}
\alias{get_genes}
\title{get_genes}
\usage{
get_genes(sample_id, product_id = FALSE, exclude_sequence = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID or A vector of sample IDs}

\item{product_id}{Optional integer product ID}

\item{exclude_sequence}{Boolean, Do not return DNA and Amino Acid sequence}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all predicted genes for a given sample.
}
\examples{
get_genes(500)
get_genes(c(500, 501, 502)
get_genes(c(500, 501), product_id=179)
}
