% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_indels}
\alias{get_indels}
\title{get_indels}
\usage{
get_indels(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all InDels present in a given sample.
}
\examples{
get_indels(500)
}
