% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-kmer.R
\name{get_kmer_partitions}
\alias{get_kmer_partitions}
\title{get_kmer_partitions}
\usage{
get_kmer_partitions(sequence, k = 31, reverse_complement = TRUE)
}
\value{
A list of partitions.
}
\description{
Get partitions in which kmers are split into.
}
