% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{get_kmers_by_sequence}
\alias{get_kmers_by_sequence}
\title{get_kmers_by_sequence}
\usage{
get_kmers_by_sequence(sample_ids, sequence, reverse_complement = TRUE)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}

\item{sequence}{A DNA sequence}

\item{reverse_complement}{Option to reverse complement k-mers or not.}
}
\value{
Parsed JSON response.
}
\description{
Given a list of sample IDs and a seqeunce return 31-mers and counts.
}
\examples{
get_kmers_by_sequence(c(500,501,502), 'ATGC')
}

