% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{get_mlst_allele_matches}
\alias{get_mlst_allele_matches}
\title{get_mlst_allele_matches}
\usage{
get_mlst_allele_matches(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve MLST allele match counts for a given sample(s).
}
\examples{
get_mlst_allele_matches(500)
get_mlst_allele_matches(c(500,501))
}
