% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{get_mlst_blast_results}
\alias{get_mlst_blast_results}
\title{get_mlst_blast_results}
\usage{
get_mlst_blast_results(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve MLST BLAST results for a given sample(s).
}
\examples{
get_mlst_blast_results(500)
get_mlst_blast_results(c(500,501))
}
