% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{get_mlst_by_sample}
\alias{get_mlst_by_sample}
\title{get_mlst_by_sample}
\usage{
get_mlst_by_sample(sample_ids)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of sample IDs return the MLST (SRST2) for each sample.
}
\examples{
get_mlst_by_sample(c(500,501,502))
}
