% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_public_cgmlst_patterns}
\alias{get_public_cgmlst_patterns}
\title{get_public_cgmlst_patterns}
\usage{
get_public_cgmlst_patterns()
}
\value{
Parsed JSON response.
}
\description{
Retrieve the cgMLST pattern counts for public samples.
}
\examples{
get_public_cgmlst_patterns()
}
