% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{get_public_tags}
\alias{get_public_tags}
\title{get_public_tags}
\usage{
get_public_tags()
}
\value{
Parsed JSON response.
}
\description{
Retrieve all public Tags in Staphopia.
}
\examples{
get_public_tags()
}
