% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_publication_links}
\alias{get_publication_links}
\title{get_publication_links}
\usage{
get_publication_links()
}
\value{
Parsed JSON response.
}
\description{
Retrieve counts for how public samples were linked to publications.
}
\examples{
get_publication_links()
}
