% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{get_publications}
\alias{get_publications}
\title{get_publications}
\usage{
get_publications(sample_id)
}
\arguments{
\item{pmid}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all publications associated with a given set of Sample ID(s).
}
\examples{
get_publications(500)
get_publications(c(500,501))
}
