% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_rank_by_year}
\alias{get_rank_by_year}
\title{get_rank_by_year}
\usage{
get_rank_by_year()
}
\value{
Parsed JSON response.
}
\description{
Retrieve the rank (Gold Silver Bronze) of public submissions by year.
}
\examples{
get_rank_by_year()
}
