% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_request}
\alias{get_request}
\title{get_request}
\usage{
get_request(url)
}
\arguments{
\item{url}{}
}
\value{
Parsed JSON response.
}
\description{
Submit a GET query to Staphopia's API. This function should not be directly
called by the user.
}

