% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{get_resistance_by_samples}
\alias{get_resistance_by_samples}
\title{get_resistance_by_samples}
\usage{
get_resistance_by_samples(sample_ids, resistance_id = NULL)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of sample IDs return the resistance phenotypes of each sample.
}
\examples{
get_resistance_by_samples(c(500,501,502))
get_resistance_by_samples(c(500,501,502), resistance_id = 1)
}

