% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{get_resistance_report}
\alias{get_resistance_report}
\title{get_resistance_report}
\usage{
get_resistance_report(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve resistances as reported by Ariba for a given sample.
}
\examples{
get_resistance_report(500)
get_resistance_report(c(500, 1000, 1500))
}
