% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_sample_by_name}
\alias{get_sample_by_name}
\title{get_sample_by_name}
\usage{
get_sample_by_name(name)
}
\arguments{
\item{name}{An string name}
}
\value{
Parsed JSON response.
}
\description{
Retrieve a sample by its name.
}
\examples{
get_sample_by_name("ERX389233")
}
