% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(sample_id)
}
\arguments{
\item{sample_id}{A Sample ID}
}
\value{
Parsed JSON response.
}
\description{
Given a sample ID(s) return sample information.
}
\examples{
get_samples(5000)
get_samples(c(5000, 5001))
}
