% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{get_samples_bt_st}
\alias{get_samples_bt_st}
\title{get_samples_by_st}
\usage{
get_samples_bt_st(st)
}
\arguments{
\item{st}{A integer MLST}
}
\value{
Parsed JSON response.
}
\description{
Given a ST, return the sample IDs with MLST (SRST2).
}
\examples{
get_samples_by_st(8)
}
