% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_samples_by_indel}
\alias{get_samples_by_indel}
\title{get_samples_by_indel}
\usage{
get_samples_by_indel(indel_id)
}
\arguments{
\item{indel_id}{An InDel ID}
}
\value{
Parsed JSON response.
}
\description{
Given a InDel ID return the samples in which InDel is present.
}
\examples{
get_samples_by_indel(c(5000))
}
