% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{get_samples_by_pmid}
\alias{get_samples_by_pmid}
\title{get_samples_by_pmid}
\usage{
get_samples_by_pmid(pmid)
}
\arguments{
\item{pmid}{An integer PubMed ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all samples associated with a given PubMed ID.
}
\examples{
get_samples_by_pmid(15155238)
}
