% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_samples_by_snp}
\alias{get_samples_by_snp}
\title{get_samples_by_snp}
\usage{
get_samples_by_snp(snp_id)
}
\arguments{
\item{snp_id}{A snp ID}
}
\value{
Parsed JSON response.
}
\description{
Given a list of snp id return the samples in which snp is present.  REWORK QUERY
}
\examples{
get_samples_by_snp(c(500,501,502))
}
