% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{get_samples_by_tag}
\alias{get_samples_by_tag}
\title{get_samples_by_tag}
\usage{
get_samples_by_tag(tag_id)
}
\arguments{
\item{tag_id}{An integer tag ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all samples associated with a given tag.
}
\examples{
get_samples_by_tag(5)
}
