% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_annotation}
\alias{get_sccmec_annotation}
\title{get_sccmec_annotation}
\usage{
get_sccmec_annotation(sample_id, exclude_sequence = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve SCCmec related results as predicted by PROKKA.
}
\examples{
get_sccmec_annotation(500)
get_sccmec_annotation(c(500, 1000, 1500))
}
