% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_ariba}
\alias{get_sccmec_ariba}
\title{get_sccmec_ariba}
\usage{
get_sccmec_ariba(sample_id, summary = FALSE, resistance_report = FALSE,
  cluster_report = FALSE, include_all = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{summary}{A boolean to return a summary of the ariba results}

\item{resistance_report}{A boolean to return a resistance report based on Ariba summary}

\item{cluster_report}{A boolean to return a cluster report based on Ariba summary}

\item{include_all}{Return all return summary results whether or not a match exists}
}
\value{
Parsed JSON response.
}
\description{
Retrieve SCCmec related results as reported by Ariba for a given sample.
}
\examples{
get_sccmec_ariba(500)
get_sccmec_ariba(c(500, 1000, 1500))
}
