% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_cassette_coverages}
\alias{get_sccmec_cassette_coverages}
\title{get_sccmec_cassette_coverages}
\usage{
get_sccmec_cassette_coverages(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve SCCmec Protein hits for a given sample(s).
}
\examples{
get_sccmec_cassette_coverages(500)
get_sccmec_cassette_coverages(c(500, 1000, 1500))
}
