% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_primers}
\alias{get_sccmec_primers}
\title{get_sccmec_primers}
\usage{
get_sccmec_primers(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve sccmec type based on primer hits for a given sample.
}
\examples{
get_sccmec_primers(500)
}
