% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_protein_hits}
\alias{get_sccmec_protein_hits}
\title{get_sccmec_protein_hits}
\usage{
get_sccmec_protein_hits(sample_id, exact_hits = FALSE, predict = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve SCCmec Protein hits for a given sample(s).
}
\examples{
get_sccmec_protein_hits(500)
get_sccmec_protein_hits(c(500, 1000, 1500))
}
