% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_subtype_hits}
\alias{get_sccmec_subtype_hits}
\title{get_sccmec_subtype_hits}
\usage{
get_sccmec_subtype_hits(sample_id, exact_hits = FALSE, predict = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{exact_hits}{Return only those hits that are an exact match. BOOL (Default FALSE)}

\item{predict}{Return a predicted SCCmec sub-type based on primer hits. BOOL (Default FALSE)}
}
\value{
Parsed JSON response.
}
\description{
Retrieve SCCmec SubType Primer hits for a given sample(s).
}
\examples{
get_sccmec_subtype_hits(500)
get_sccmec_subtype_hits(500, exact_hits=TRUE)
get_sccmec_subtype_hits(500, predict=TRUE)
get_sccmec_subtype_hits(c(500, 1000, 1500), predict=TRUE)
}
