% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_subtypes}
\alias{get_sccmec_subtypes}
\title{get_sccmec_subtypes}
\usage{
get_sccmec_subtypes(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID or A vector of sample IDs}

\item{product_id}{Optional integer product ID}

\item{cluster_id}{Optional integer cluster ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all predicted sccmec subtypes for a given sample.
}
\examples{
get_sccmec_subtypes(500)
get_sccmec_subtypes(c(500, 501, 502)
get_sccmec_subtypes(c(500, 501)))
}
