% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlst.R
\name{get_sequence_type}
\alias{get_sequence_type}
\title{get_sequence_type}
\usage{
get_sequence_type(sample_id, blast = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{blast}{Get blast based results for each Loci, not SRST2. BOOL (Default: FALSE)}
}
\value{
Parsed JSON response.
}
\description{
Retrieve MLST results based on SRST2 for a given sample(s).
}
\examples{
get_sequence_type(500)
get_sequence_type(c(500,501))
}
