% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_sequencing_stats_by_year}
\alias{get_sequencing_stats_by_year}
\title{get_sequencing_stats_by_year}
\usage{
get_sequencing_stats_by_year(is_original = FALSE)
}
\arguments{
\item{is_original}{Boolean}
}
\value{
Parsed JSON response.
}
\description{
Retrieve sequencing stats by year based on when samples were first made public.
}
\examples{
get_sequencing_stats_by_year(is_original=TRUE)
}
