% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_snp_by_position}
\alias{get_snp_by_position}
\title{get_snp_by_position}
\usage{
get_snp_by_position(position)
}
\arguments{
\item{position}{An integer reference position}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all SNP IDs for a given reference position.
}
\examples{
get_snp_by_position(500)
}
