% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_snps_by_samples}
\alias{get_snps_by_samples}
\title{get_snps_by_samples}
\usage{
get_snps_by_samples(sample_ids)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of sample IDs return the SNPs present in each sample.
}
\examples{
get_snps_by_samples(c(500,501,502))
}

