% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_snps_in_bulk}
\alias{get_snps_in_bulk}
\title{get_snps_in_bulk}
\usage{
get_snps_in_bulk(snp_ids)
}
\arguments{
\item{snp_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of SNP IDs return information about each SNP.
}
\examples{
get_snps_in_bulk(c(5000,5001,5002))
}
