% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_st}
\alias{get_st}
\title{get_st}
\usage{
get_st(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve MLST results based on SRST2 for a given sample.
}
\examples{
get_st(500)
}
