% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_st_blast}
\alias{get_st_blast}
\title{get_st_blast}
\usage{
get_st_blast(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve MLST results based on BLAST hits for a given sample.
}
\examples{
get_st_blast(500)
}

