% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_st_by_year}
\alias{get_st_by_year}
\title{get_st_by_year}
\usage{
get_st_by_year()
}
\value{
Parsed JSON response.
}
\description{
Retrieve the sequence type assignments of public submissions by year.
}
\examples{
get_st_by_year()
}
