% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{get_tag_by_name}
\alias{get_tag_by_name}
\title{get_tag_by_name}
\usage{
get_tag_by_name(tag)
}
\arguments{
\item{tag}{The tag name as a string}
}
\value{
Parsed JSON response.
}
\description{
Retrieve retrieve a tag by its name.
}
\examples{
get_tag_by_name('my-tag-name')
}
