% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_tags}
\alias{get_tags}
\title{get_tags}
\usage{
get_tags(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all Tags associated with a given sample.
}
\examples{
get_tags(500)
}

