% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_top_sequence_types}
\alias{get_top_sequence_types}
\title{get_top_sequence_types}
\usage{
get_top_sequence_types(total = 10)
}
\arguments{
\item{total}{Integer}
}
\value{
Parsed JSON response.
}
\description{
Retrieve top X (default 10) sequence sequence types.
}
\examples{
get_top_sequence_types(20)
}
