% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variant_annotation}
\alias{get_variant_annotation}
\title{get_variant_annotation}
\usage{
get_variant_annotation(id = NULL, locus_tag = NULL)
}
\arguments{
\item{id}{An integer annotation ID}

\item{locus_tag}{An string locus_tag to filter annotaions on}
}
\value{
Parsed JSON response.
}
\description{
Get annotation info associated with a variant.
}
\examples{
get_variant_annotation(7)
get_variant_annotation(locus_tag="SA_RS00145")
}
