% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variant_count_by_position}
\alias{get_variant_count_by_position}
\title{get_variant_count_by_position}
\usage{
get_variant_count_by_position(ids = FALSE, is_annotation = FALSE)
}
\arguments{
\item{ids}{A vector of positions or Annotation IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of positions or annotation ids return SNP/InDel counts.
}
\examples{
get_variant_count_by_position(c(1,2,3))
get_variant_count_by_position(c(1,2,3), is_annotation = TRUE)
}
