% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variant_counts}
\alias{get_variant_counts}
\title{get_variant_counts}
\usage{
get_variant_counts(sample_id)
}
\arguments{
\item{sample_id}{A vector of Sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of Sample IDs return SNP/InDel counts.
}
\examples{
get_variant_counts(500)
get_variant_counts(c(500,5002,4003))
}
