% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variant_gene_sequence}
\alias{get_variant_gene_sequence}
\title{get_variant_gene_sequence}
\usage{
get_variant_gene_sequence(sample_ids, annotation_ids, debug = FALSE)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}

\item{annotation_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of Sample IDs and Annotation IDs return the SNP substituted
gene sequences. Sequences are concatenated.
}
\examples{
get_variant_gene_sequence(c(5000, 5001, 5002), c(2, 3, 4))
}
