% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virulence.R
\name{get_virulence_results}
\alias{get_virulence_results}
\title{get_virulence_results}
\usage{
get_virulence_results(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Retrieve virulences as reported by Ariba for a given sample.
}
\examples{
get_virulence_results(500)
get_virulence_results(c(500, 1000, 1500))
}
