% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwas_lr}
\alias{gwas_lr}
\title{gwas_lr}
\usage{
gwas_lr(snps, phenotype, names)
}
\arguments{
\item{snps}{A vector of SNPs present (1) or absent (0) per sample}

\item{names}{Name of the SNPs being tested}

\item{phenotypes}{A vector of phenotypes per sample}
}
\value{
A dataframe of linear regression results.
}
\description{
Test for significance between a phenotype and SNPs. This function should not
be called by the user.'
}
