% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.not.null}
\alias{is.not.null}
\title{is.not.null}
\usage{
is.not.null(x)
}
\arguments{
\item{x}{}
}
\value{
bool TRUE is not null or FALSE if null.
}
\description{
A wrapper for "!is.null". This function should not be directly used by the
user.
}

