% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_multiple_ids}
\alias{is_multiple_ids}
\title{is_multiple_ids}
\usage{
is_multiple_ids(x)
}
\arguments{
\item{x}{}
}
\value{
bool TRUE is multiple ids else FALSE.
}
\description{
A wrapper for to validate given vector is multiple ids and proper type. This
function should not be directly used by the user.
}
