% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_single_id}
\alias{is_single_id}
\title{is_single_id}
\usage{
is_single_id(x)
}
\arguments{
\item{x}{}
}
\value{
bool TRUE is a single id else FALSE.
}
\description{
A wrapper for to validate given id is single and proper type. This function
should not be directly used by the user.
}
