% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression}
\usage{
linear_regression(phenotype, snp, position)
}
\arguments{
\item{phenotype}{A vector of phenotypes per sample}

\item{snp}{A vector of SNP present (1) or absent (0) per sample}

\item{name}{Name of the SNP being tested}
}
\value{
Results from linear regression.
}
\description{
Test for significance between a phenotype and SNP. This function should not
be called by the user.'
}

