% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plots.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{manhattan_plot}
\usage{
manhattan_plot(df, genome_size = 2814816, alpha = 0.05,
  correction = "bonferroni", xtitle = "Position along chromosome")
}
\arguments{
\item{df}{A data frame returned from 'run_gwas'
'}
}
\value{
ggplot2 plot.
}
\description{
Produce a manhattan plot of GWAS results.  CLEANUP UP
}
\examples{
manhattan_plot(df)
}
