% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-heatmap.R
\name{pangenome_heatmap}
\alias{pangenome_heatmap}
\title{pangenome_matrix}
\usage{
pangenome_heatmap(pg_matrix, main = "Pangenome Matrix Heatmap",
  xlab = "Cluster ID", ylab = "Samples")
}
\arguments{
\item{tag}{The tag name as a string}
}
\value{
A list with total genes (int), number of core genes (int), 
number of accessory genes (int), array of core genes (chr)
and array of accessory genes (chr)
}
\description{
Construct a matrix with samples as rows and gene cluster IDs as columns, 
where the values are a binary indicating presence of that gene for that 
sample (1) or absence (0)
}
\examples{
my_matrix.stats<-pangenome_heatmap(my_matrix)
my_matrix.stats<-pangenome_heatmap(my_matrix,'Heatmap','Genes','Samples')
my_matrix.stats<-pangenome_heatmap(pangenome_matrix('my-tag-name'))
}
