% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-heatmap.R
\name{pangenome_matrix}
\alias{pangenome_matrix}
\title{pangenome_matrix}
\usage{
pangenome_matrix(samples)
}
\arguments{
\item{samples}{A vector of sample IDS. Can be created by get_samples_by_tag().}
}
\value{
Parsed JSON response.
}
\description{
Construct a matrix with samples as rows and gene cluster IDs as columns, 
where the values are a binary indicating presence of that gene for that 
sample (1) or absence (0).
}
\examples{
pangenome_matrix('sample_vector')
}
