% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{plasmid_meta}
\alias{plasmid_meta}
\title{plasmid_meta}
\usage{
plasmid_meta(contigs)
}
\arguments{
\item{contig_data(df)}{}
}
\value{
data.frame
}
\description{
plasmid_meta
}
\examples{
plasmid_meta(get_contigs(500))
}

