% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_request}
\alias{post_request}
\title{post_request}
\usage{
post_request(url, data)
}
\arguments{
\item{url}{}

\item{data}{A vector of data to be POSTED}
}
\value{
Parsed JSON response.
}
\description{
Submit a POST query to Staphopia's API. This function should not be directly
called by the user.
}
