% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qq_plot}
\alias{qq_plot}
\title{qq_plot}
\usage{
qq_plot(df, alpha = 0.05, correction = "bonferroni")
}
\arguments{
\item{df}{A data frame returned from 'run_gwas'
'}
}
\value{
ggplot2 plot.
}
\description{
Produce a QQ plot of GWAS results.
}
\examples{
qq_plot(df)
}
