% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{rev_comp}
\alias{rev_comp}
\title{rev_comp}
\usage{
rev_comp(sequence)
}
\arguments{
\item{sequence}{A DNA sequence}
}
\value{
Reverse complemented sequence as a string.
}
\description{
Reverse complement an input sequence.
}
