% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{run_gwas}
\alias{run_gwas}
\title{run_gwas}
\usage{
run_gwas(snp_matrix, phenotype, mode = "lr")
}
\arguments{
\item{snp_matrix}{A vector of SNPs present (1) or absent (0) per sample}

\item{mode}{GWAS test to run.}

\item{mic}{A vector of phenotypes per sample}
}
\value{
Results from GWAS
}
\description{
Given a list of sample IDs return the resistance phenotypes of each sample.
}
\examples{
run_gwas(snp_matrix, phenotype)
}

