% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{samp_clust}
\alias{samp_clust}
\title{samp_clust}
\usage{
samp_clust(gene.list)
}
\arguments{
\item{gene.list}{Parsed JSON response from the get_genes function.}
}
\value{
matrix.
}
\description{
In case scientific notation is turned on, convert it to an integer. This
function should not be directly used by the user.
}

