% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{split_sequence_into_kmers}
\alias{split_sequence_into_kmers}
\title{split_sequence_into_kmers}
\usage{
split_sequence_into_kmers(sequence, k = 31, reverse_complement = TRUE)
}
\arguments{
\item{sequence}{A DNA sequence}

\item{k}{A value of 'k' to split sequence into. Default is 31.}

\item{reverse_complement}{Option to reverse complement k-mers or not.}
}
\value{
A vecter of k-mers.
}
\description{
Split a seqeunce into a set of k-mers.
}

