% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_vector_into_chunks}
\alias{split_vector_into_chunks}
\title{split_vector}
\usage{
split_vector_into_chunks(x, size)
}
\arguments{
\item{x}{Vector to be split.}

\item{size}{A number elements to split a vector by.}
}
\value{
bool TRUE is not null or FALSE if null.
}
\description{
Split a vector into smaller vectors based on the given size. This function
should not be directly used by the user.
}

