% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_get_request}
\alias{submit_get_request}
\title{submit_get_request}
\usage{
submit_get_request(request)
}
\arguments{
\item{request}{}
}
\value{
Parsed JSON response.
}
\description{
Prepare a GET query for submission to Staphopia's API. Paginated responses
are aggragated and submitted as a single response. This function should
not be directly called by the user.
}

