% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_paginated_request}
\alias{submit_paginated_request}
\title{submit_paginated_request}
\usage{
submit_paginated_request(next_page)
}
\arguments{
\item{next_page}{}
}
\value{
Parsed JSON response.
}
\description{
Automates cycling through all pages of a paginated response. A delay of
200ms is introduced between each query so as to not overload the API. This
function should not be directly called by the user.
}
