% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_post_request}
\alias{submit_post_request}
\title{submit_post_request}
\usage{
submit_post_request(request, data)
}
\arguments{
\item{request}{}
}
\value{
Parsed JSON response.
}
\description{
Prepare a POST query for submission to Staphopia's API. Paginated responses
are aggregated and submitted as a single response. This function should
not be directly called by the user.
}

