% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_staphopia}
\alias{test_staphopia}
\title{get_resistance_results}
\usage{
test_staphopia()
}
\value{
Parsed JSON response.
}
\description{
Retrieve resistances as reported by Ariba for a given sample.
}
\examples{
test_staphopia()
}
