% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_assembly}
\alias{write_assembly}
\title{write_assembly}
\usage{
write_assembly(sample_id, output_dir = "./", plasmids = FALSE,
  coverage = 2)
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{output_dir}{A string, directory to output fasta file(s) to. Default: ./}

\item{plasmids}{Return plasmid stats. BOOL (Default FALSE)}

\item{coverage}{A numeric value. All contigs with coverage lower than this are removed. Default: 2}
}
\value{
write_assembly
}
\description{
Write the assembled contigs to a FASTA file for a given sample.
}
\details{
Example headers:
>500|ERX105420|1 coverage=38.394969 length=225174 analysis_version=112017
>500|ERX105420|1|plasmid coverage=50.715860 length=41605 analysis_version=112017
}
\examples{
write_assembly(500,"./protein_seqs/"
output fasta header is of the format <sample>_<staphopia proteinid>_<product id>_<cluster id>_productname
}
