% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_contigs_to_fasta}
\alias{write_contigs_to_fasta}
\title{write_contigs_to_fasta}
\usage{
write_contigs_to_fasta(sample_id, fasta_directory, filt_contigs = 2)
}
\arguments{
\item{sample_id}{(numeric)}

\item{fasta_directory}{(string)}

\item{filt_contigs}{(numeric) (all contigs with coverage lower than this are removed)}
}
\value{
none
}
\description{
write_contigs_to_fasta
}

