% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_genes}
\alias{write_genes}
\title{write_genes}
\usage{
write_genes(sample_id, output_dir = "./", include_rna = TRUE)
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{output_dir}{A string, directory to output fasta file(s) to. Default: ./}

\item{include_rna}{Include RNAs in the output (Default TRUE)}
}
\value{
write_genes
}
\description{
Write the predicted genes (DNA) to a FASTA file for a given sample.
}
\details{
Example headers:
>500|ERX105420_00001|1678|YP_005744112.1 [product=histidinol-phosphate aminotransferase] [name=none] [note=none]
500|ERX105420_00001|1678|YP_005744112.1 ---> sample_id|locus_tag|product_id|inference
}
\examples{
write_genes(500)
write_genes(500, include_rna = FALSE)
}
