% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_plasmids_to_fasta}
\alias{write_plasmids_to_fasta}
\title{write_plasmids_to_fasta}
\usage{
write_plasmids_to_fasta(sample_id, output_dir = "./", filt_contigs = 2)
}
\arguments{
\item{sample_id}{(numeric)}

\item{output_dir}{(string) (ending in '/')}

\item{filt_contigs}{(numeric) (all contigs with coverage lower than this are removed; default = 2)}
}
\value{
(data.frame) (summary of combined plasmid components)
}
\description{
write_plasmids_to_fasta
}
\examples{
write_plasmids_to_fasta(500, output_dir='~/')
}
