% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_plot}
\alias{write_plot}
\title{write_plot}
\usage{
write_plot(plot_object, name, height = 5, width = 12)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{name}{Basename for the output PDF and PNG files}

\item{height}{The PDF height of the output (Default: 5)}

\item{width}{The PDF width of the object (Default: 12)}
}
\value{
bool TRUE is multiple ids else FALSE.
}
\description{
A wrapper for to validate given vector is multiple ids and proper type. This
function should not be directly used by the user.
}
