% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_proteins}
\alias{write_proteins}
\title{write_proteins}
\usage{
write_proteins(sample_id, output_dir = "./")
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{output_dir}{A string, directory to output fasta file(s) to. Default: ./}
}
\value{
write_proteins
}
\description{
Write the predicted proteins (AA) to a FASTA file for a given sample.
}
\details{
Example headers:
>500|ERX105420_00001|1678|YP_005744112.1 [product=histidinol-phosphate aminotransferase] [name=none] [note=none]
500|ERX105420_00001|1678|YP_005744112.1 ---> sample_id|locus_tag|product_id|inference
}
\examples{
write_proteins(500)
}
