% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_proteins_2_fasta}
\alias{write_proteins_2_fasta}
\title{write_proteins_2_fasta}
\usage{
write_proteins_2_fasta(sample, output_dir = "./")
}
\arguments{
\item{sample}{(single sample_id)}

\item{output_dir}{(with following '/', e.g: "./myproteins/")}
}
\value{
write_proteins_2_fasta
}
\description{
write_proteins_2_fasta
}
\examples{
write_proteins_2_fasta(500,"./protein_seqs/"
output fasta header is of the format <sample>_<staphopia proteinid>_<product id>_<cluster id>_productname
}
