% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{get_assembly_contigs}
\alias{get_assembly_contigs}
\title{get_assembly_contigs}
\usage{
get_assembly_contigs(sample_id, plasmids = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{plasmids}{Return plasmid stats. BOOL (Default FALSE)}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all assembled contigs for a given sample(s).
}
\examples{
get_assembly_contigs(500)
get_assembly_contigs(500, plasmids=TRUE)
get_assembly_contigs(c(500, 1000, 1500))
}
