% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{get_assembly_stats}
\alias{get_assembly_stats}
\title{get_assembly_stats}
\usage{
get_assembly_stats(sample_id, plasmids = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{plasmids}{Return plasmid stats. BOOL (Default FALSE)}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all stats associated with an assembly for a given sample.
}
\examples{
get_assembly_stats(500)
get_assembly_stats(500, plasmids=TRUE)
get_assembly_stats(c(500, 1000, 1500))
}
