% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_gene_product}
\alias{get_gene_product}
\title{get_gene_product}
\usage{
get_gene_product(product_id)
}
\arguments{
\item{product_id}{An integer product ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve gene product infor for a given ID.
}
\examples{
get_gene_product(500)
}
