% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_gene_products}
\alias{get_gene_products}
\title{get_gene_products}
\usage{
get_gene_products(term = FALSE)
}
\value{
Parsed JSON response.
}
\description{
Retrieve all gene products in Staphopia.
}
\examples{
get_gene_products()
}
