% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_genes}
\alias{get_genes}
\title{get_genes}
\usage{
get_genes(sample_id, product_id = NULL)
}
\arguments{
\item{sample_id}{An integer sample ID}

\item{product_id}{Optional integer product ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all predicted genes for a given sample.
}
\examples{
get_genes(500)
}

