% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_indels}
\alias{get_indels}
\title{get_indels}
\usage{
get_indels(sample_id, annotation_id = NULL)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{annotation_id}{Filter results based on an annotation ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all InDels present in a given sample(s).
}
\examples{
get_indels(500)
get_indels(c(500,501))
}
