% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_public_samples}
\alias{get_public_samples}
\title{get_public_samples}
\usage{
get_public_samples(include_location = FALSE, limit = FALSE)
}
\arguments{
\item{include_location}{A boolean, includes location information with samples}

\item{limit}{An integer, limit the number of samples returned}
}
\value{
Parsed JSON response.
}
\description{
Retrieve publicly available ENA samples.
}
\examples{
get_public_samples()
get_public_samples(include_location = TRUE)
get_public_samples(include_location = TRUE, limit = 5)
}
