% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_published_samples}
\alias{get_published_samples}
\title{get_published_samples}
\usage{
get_published_samples()
}
\value{
Parsed JSON response.
}
\description{
Retrieve published ENA samples.
}
\examples{
get_published_samples()
}

