% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{get_resistance}
\alias{get_resistance}
\title{get_resistance}
\usage{
get_resistance(antibiotic = NULL, test = NULL)
}
\arguments{
\item{antibiotic}{An antibiotic to search for}

\item{test}{An antibiotic test method to search for.}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all resistances in the database. Can be filtered based on
antibiotic and test.
}
\examples{
get_resistance()
get_resistance(antibiotic="Vancomycin")
get_resistance(antibiotic="Vancomycin", test="Etest")
}

