% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{get_resistance_results}
\alias{get_resistance_results}
\title{get_resistance_results}
\usage{
get_resistance_results(sample_id, summary = FALSE,
  resistance_report = FALSE, cluster_report = FALSE)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{summary}{A boolean to return a summary of the ariba results}

\item{resistance_report}{A boolean to return a resistance report based on Ariba summary}

\item{cluster_report}{A boolean to return a cluster report based on Ariba summary}
}
\value{
Parsed JSON response.
}
\description{
Retrieve resistances as reported by Ariba for a given sample.
}
\examples{
get_resistance_results(500)
get_resistance_results(c(500, 1000, 1500))
}
