% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_samples_by_snp}
\alias{get_samples_by_snp}
\title{get_samples_by_snp}
\usage{
get_samples_by_snp(snp_id)
}
\arguments{
\item{snp_id}{An SNP ID}
}
\value{
Parsed JSON response.
}
\description{
Given a SNP ID(s) return the samples in which SNP is present.
}
\examples{
get_samples_by_snp(c(5000))
}
