% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_samples_by_snp}
\alias{get_samples_by_snp}
\title{get_samples_by_snp}
\usage{
get_samples_by_snp(snp_id)
}
\arguments{
\item{snp_id}{A snp ID}
}
\value{
Parsed JSON response.
}
\description{
Given a list of snp id return the samples in which snp is present.
}
\examples{
get_snps_by_sample(c(500,501,502))
}

