% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccmec.R
\name{get_sccmec_subtype}
\alias{get_sccmec_subtype}
\title{get_sccmec_subtype}
\usage{
get_sccmec_subtype(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve sccmec subtype based on primer hits for a given sample.
}
\examples{
get_sccmec_subtype(500)
}

