% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_sequence_quality}
\alias{get_sequence_quality}
\title{get_sequence_quality}
\usage{
get_sequence_quality(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve sequence quality metrics for a given sample.
}
\examples{
get_sequence_quality(500)
}
