% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_snps}
\alias{get_snps}
\title{get_snps}
\usage{
get_snps(sample_id, annotation_id = NULL)
}
\arguments{
\item{sample_id}{An integer sample ID, or vector of sample IDs}

\item{annotation_id}{An integer annotation ID to filter snps on}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all SNPs present in a given sample(s).
}
\examples{
get_snps(500)
get_snps(c(500,501))
}
