% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_snps}
\alias{get_snps}
\title{get_snps}
\usage{
get_snps(sample_id)
}
\arguments{
\item{sample_id}{An integer sample ID}
}
\value{
Parsed JSON response.
}
\description{
Retrieve all SNPs present in a given sample.
}
\examples{
get_snps(500)
}
