% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_submission_by_year}
\alias{get_submission_by_year}
\title{get_submission_by_year}
\usage{
get_submission_by_year(all = FALSE)
}
\arguments{
\item{all}{Boolean, include all available sample, not just the processed ones}
}
\value{
Parsed JSON response.
}
\description{
Retrieve the public submissions by year.
}
\examples{
get_submission_by_year()
}
