% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stagnant-functions.R
\name{get_variant_counts}
\alias{get_variant_counts}
\title{get_variant_counts}
\usage{
get_variant_counts(sample_ids)
}
\arguments{
\item{sample_ids}{A vector of sample IDs}
}
\value{
Parsed JSON response.
}
\description{
Given a list of Sample IDs return SNP/InDel counts.  REWORK
}
\examples{
get_variant_counts(c(5000,5001,5002))
}
