% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{pangenome_matrix}
\alias{pangenome_matrix}
\title{pangenome_matrix}
\usage{
pangenome_matrix(tag)
}
\arguments{
\item{tag}{The tag name as a string}
}
\value{
Parsed JSON response.
}
\description{
Construct a matrix with samples as rows and gene cluster IDs as columns, 
where the values are a binary indicating presence of that gene for that 
sample (1) or absence (0).
}
\examples{
pangenome_matrix('my-tag-name')
}

