% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_post_request}
\alias{submit_post_request}
\title{submit_post_request}
\usage{
submit_post_request(request, data, chunk_size = 10)
}
\arguments{
\item{request}{An API endpoint
'}

\item{data}{A vector of data to be POSTED
'}

\item{chunk_size}{Optional parameter to determine size of chunks}
}
\value{
Parsed JSON response.
}
\description{
Prepare a POST query for submission to Staphopia's API. By default post
requests will be split into a number of smaller requests. This is hopefully
avoid timeouts.
}

