% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{write_contigs_to_fasta}
\alias{write_contigs_to_fasta}
\title{write_contigs_to_fasta}
\usage{
write_contigs_to_fasta(sample_id, output_dir = "./", filt_contigs = 2)
}
\arguments{
\item{sample_id}{(numeric)}

\item{filt_contigs}{(numeric) (all contigs with coverage lower than this are removed)}

\item{fasta_directory}{(string) (ending in '/')}
}
\description{
write_contigs_to_fasta
}
\examples{
write_contigs_to_fasta(500)
write_contigs_to_fasta(500, output_dir='~/')
}
